Inline NoConstructorAgrumentsForInterfaceMock_2
