+++ Source/Moq.Analyzers.Test/NoConstructorArgumentsForInterfaceMockAnalyzerTests.cs
@@ -99,20 +99,142 @@ public class NoConstructorArgumentsForInterfaceMockAnalyzerTests : DiagnosticVer
     [Fact]
     public Task ShouldPassIfCustomMockClassIsUsed()
     {
-        return Verify(VerifyCSharpDiagnostic(File.ReadAllText("Data/NoConstructorArgumentsForInterfaceMock.TestFakeMoq.cs")));
+        return Verify(VerifyCSharpDiagnostic(
+            [
+                """
+                namespace NoConstructorArgumentsForInterfaceMock.TestFakeMoq;
+
+                public enum MockBehavior
+                {
+                    Default,
+                    Strict,
+                    Loose,
+                }
+
+                internal interface IMyService
+                {
+                    void Do(string s);
+                }
+
+                public class Mock<T>
+                    where T : class
+                {
+                    public Mock() { }
+
+                    public Mock(params object[] ar) { }
+
+                    public Mock(MockBehavior behavior) { }
+
+                    public Mock(MockBehavior behavior, params object[] args) { }
+                }
+
+                internal class MyUnitTests
+                {
+                    private void TestFakeMoq()
+                    {
+                        var mock1 = new Mock<IMyService>("4");
+                        var mock2 = new Mock<IMyService>(5, true);
+                        var mock3 = new Mock<IMyService>(MockBehavior.Strict, 6, true);
+                        var mock4 = new Mock<IMyService>(Moq.MockBehavior.Default, "5");
+                        var mock5 = new Mock<IMyService>(MockBehavior.Strict);
+                        var mock6 = new Mock<IMyService>(MockBehavior.Loose);
+                    }
+                }
+                """
+            ]));
     }
 
     // TODO: This feels duplicated with other tests
     [Fact]
     public Task ShouldFailIsRealMoqIsUsedWithInvalidParameters()
     {
-        return Verify(VerifyCSharpDiagnostic(File.ReadAllText("Data/NoConstructorArgumentsForInterfaceMock.TestRealMoqWithBadParameters.cs")));
+        return Verify(VerifyCSharpDiagnostic(
+            [
+                """
+                namespace NoConstructorArgumentsForInterfaceMock.TestRealMoqWithBadParameters;
+
+                public enum MockBehavior
+                {
+                    Default,
+                    Strict,
+                    Loose,
+                }
+
+                internal interface IMyService
+                {
+                    void Do(string s);
+                }
+
+                public class Mock<T>
+                    where T : class
+                {
+                    public Mock() { }
+
+                    public Mock(params object[] ar) { }
+
+                    public Mock(MockBehavior behavior) { }
+
+                    public Mock(MockBehavior behavior, params object[] args) { }
+                }
+
+                internal class MyUnitTests
+                {
+                    private void TestRealMoqWithBadParameters()
+                    {
+                        var mock1 = new Moq.Mock<IMyService>(1, true);
+                        var mock2 = new Moq.Mock<IMyService>("2");
+                        var mock3 = new Moq.Mock<IMyService>(Moq.MockBehavior.Default, "3");
+                        var mock4 = new Moq.Mock<IMyService>(MockBehavior.Loose, 4, true);
+                        var mock5 = new Moq.Mock<IMyService>(MockBehavior.Default);
+                        var mock6 = new Moq.Mock<IMyService>(MockBehavior.Default);
+                    }
+                }
+                """
+            ]));
     }
 
     [Fact]
     public Task ShouldPassIfRealMoqIsUsedWithValidParameters()
     {
-        return Verify(VerifyCSharpDiagnostic(File.ReadAllText("Data/NoConstructorArgumentsForInterfaceMock.TestRealMoqWithGoodParameters.cs")));
+        return Verify(VerifyCSharpDiagnostic(
+            [
+                """
+                namespace NoConstructorArgumentsForInterfaceMock.TestRealMoqWithGoodParameters;
+
+                public enum MockBehavior
+                {
+                    Default,
+                    Strict,
+                    Loose,
+                }
+
+                internal interface IMyService
+                {
+                    void Do(string s);
+                }
+
+                public class Mock<T>
+                    where T : class
+                {
+                    public Mock() { }
+
+                    public Mock(params object[] ar) { }
+
+                    public Mock(MockBehavior behavior) { }
+
+                    public Mock(MockBehavior behavior, params object[] args) { }
+                }
+
+                internal class MyUnitTests
+                {
+                    private void TestRealMoqWithGoodParameters()
+                    {
+                        var mock1 = new Moq.Mock<IMyService>(Moq.MockBehavior.Default);
+                        var mock2 = new Moq.Mock<IMyService>(Moq.MockBehavior.Default);
+                    }
+                }
+                """
+            ]));
     }
 
     protected override DiagnosticAnalyzer GetCSharpDiagnosticAnalyzer()
